refactor(exec): move PATH lookup to a leaf, fix the exec-to-registry inversion
exec/run.ts, the lowest-level subprocess primitive, imported isOnPath from plugins/registry.ts (which imports every plugin): the lowest layer reaching up into the highest. Move pathTo/isOnPath to a new leaf module src/exec/on-path.ts (fs/path only); run.ts imports it directly and the registry re-exports it for its consumers (doctor report, buildRegistry).

The ExecRunner interface is kept as run/runJson/onPath. Narrowing it (architecture review candidate 6) was assessed and deferred: dropping runJson churns four impls plus six test files for two callers, and dropping onPath reintroduces a defaults-registry cycle. Recorded in ADR 0032.
